// game/src/game/Models/Player/enums.ts
export enum DirectionEnum {
  west = 'west',
  northWest = 'northWest',
  north = 'north',
  northEast = 'northEast',
  east = 'east',
  southEast = 'southEast',
  south = 'south',
  southWest = 'southWest',
}

export enum MotionEnum {
  idle = 'idle',
  walk = 'walk',
  attack = 'attack',
  die = 'die',
  shoot = 'shoot',
}